+++ activerecord/test/schema/schema.rb
@@ -877,6 +877,13 @@ ActiveRecord::Schema.define do
     t.string :overloaded_string_with_limit, limit: 255
     t.string :string_with_default, default: 'the original default'
   end
+
+  create_table :users, force: true do |t|
+    t.string :token
+    t.string :name
+  end
+
+  add_index :users, :token, unique: true, name: 'unique_token_index'
 end
 
 Course.connection.create_table :courses, force: true do |t|